src/requirements.txt
alembic==1.8.0
anyio==3.6.1
APScheduler==3.9.1
asgiref==3.5.2
bcrypt==3.2.2
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
dnspython==2.2.1
email-validator==1.2.1
fastapi==0.78.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
httptools==0.4.0
idna==3.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
multipledispatch==0.6.0
orjson==3.7.6
psycopg2-binary==2.9.3
pycparser==2.21
pydantic==1.9.1
pydig==0.4.0
PyJWT==2.4.0
python-crontab==2.6.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
python-nmap==0.7.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.3.0
tzdata==2022.1
tzlocal==4.2
ujson==5.4.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
Werkzeug==2.1.2
